Give a fix to the turret bug
I think that the turret bug was because I incremented or decremented the
turret speed based on key press/release. If the key release happens to
not have registered when the spacebar is pressed, or just at any time,
then there's no way to stop the turret from moving. This proposes a fix
for Tank turret stops accepting input, stuck high or low #15, by setting
the speed on key press and zeroing it on key release.