Cgrooves/issue9 (#14)
* First pass make tank with turret

Controller sucks, but hey, it draws a "tank", and
you can change the turret angle.

* A slightly improved keyboard interface

* Separate tank out into module

* Add some explanatory comments

* Escape key exits program

* A potential fix, maybe not

Not sure if this math is right

* Controllable tank angle
Fixes #5

* Implement simple text display for angle

There's probably a better way to implement this using the
`arcade.gui.UILabel` module, but this is fine for now.

* Change sky color to something a little darker

* Add another tank

Pretty ugly coding, but just adds in the second tank
in order to close out Player tanks on main window #4

* Make a Player list, put shapes list in setup

The shapes list needed to be in setup, and make the players into
a list instead of hard-coded each one.

* Add switching logic on spacebar

* Make num of tanks configurable

* Switch to using a cycle for cycling tanks

The itertools.cycle is neater than the logic I implemented,
so switch over to that.

* Give a fix to the turret bug

I think that the turret bug was because I incremented or decremented the
turret speed based on key press/release. If the key release happens to
not have registered when the spacebar is pressed, or just at any time,
then there's no way to stop the turret from moving. This proposes a fix
for Tank turret stops accepting input, stuck high or low #15, by setting
the speed on key press and zeroing it on key release.

* Add more explanatory "why" comments

* Simplify import statements

Since a huge part of this project is for learning, I think that it's a
good idea to just keep things simple by making a rule for ourselves
that, unless something is a super long use statements (i.e. if we're
having to type somepackage.foobar.baz.bin.AwesomeFunction, which would
warrant a statements like "from somepackage.foobar.baz.bin import
AwesomeFunction" so that we don't have to type out the long thing every
single time we want to use AwesomeFunction), just do a simple "import
<package>" statement. That way, its clear where things are coming from.

* Add some more clarifying comments to Tank