Basic Firing (#16)
* Add power control to tank

* Add basic physics-based bullet firing

No collision detection yet, or anything fancy, but this adds the
functionality that when a tank fires, it actually fires a bullet, which
is passed into the physics engine. Many tweaks need to be made.

* Add a global constants file

I think eventually we'll want this to be a config file, but for now,
just to keep things simple, add constants that multiple modules will
need, like the window size, to here.

* Turn ground into a sprite and add physics

In order to make it easier to do physics and collisions with the ground,
construct it as a rectangular sprite instead of just a shape.

* Make bullet disappear when it leaves the screen

If the bullet goes off the edge of the screen, width-wise, we want it
to disappear for now.

* Add collision between bullet and ground

* Add artillery round detonation

Add a function call for detonation, which will be overloaded and used
for other types of weaponry.

* Create a weapons package

* Remove responsibility of creating bullet sprite from tank

* Create a cycling weapons list, and a HUD entry

* Create global weapons list

* Move constants into a config file

* Clean up main game logic

Clean up some long functions and make them more reasonable to read
and follow by having them call other functions with (hopefully) useful
names.

* Rename bullet to weapon

Try to introduce a little consistency here.